Move Transport Node toggle to a new Advanced settings card
The Transport Node toggle was nestled inside the Network card, which
conflated "here's how to configure interfaces" with a power-user decision
about whether this device relays mesh traffic for other peers. Users who
just wanted to add a TCP interface got prompted to consider a routing
decision they probably shouldn't touch.

Move it to a new dedicated Advanced card between RNode Flasher and About
on the Settings page — findable for power users, out of the way for
everyone else. The new card carries the same toggle with the same wiring
(SettingsViewModel.transportNodeEnabled / setTransportNodeEnabled), so
behavior and persisted state are unchanged.

Changes:
- Added SettingsCardId.ADVANCED between SHARE_COLUMBA and ABOUT.
- Created AdvancedCard composable; mirrors the layout pattern of the
  other settings cards.
- Removed the Transport Node row + associated props from NetworkCard
  (networkCard now only concerns Network Status and Manage Interfaces).
- Moved the 7 transport-node unit tests from NetworkCardTest to a new
  AdvancedCardTest. Network-card default-values test updated to drop
  the Transport Node assertion.
- Wired AdvancedCard into SettingsScreen between RNodeFlasherCard and
  AboutCard with the same viewModel state/callbacks.
